// src/utils/enums.ts
export enum UserRole {
  RECIPIENT = 'RECIPIENT',
  PROVIDER = 'PROVIDER'
}

export enum contractType {
  MONTHLY = "MONTHLY",
  QUERTERLY = "QUERTERLY",
  YEARLY = "YEARLY"
}

export enum contractStatus {
  SIGNED = "SIGNED",
  AVAILABLE = "AVAILABLE"
}

export enum CashkicksStatus {
  PENDING = "PENDING",
  APPROVED = "APPROVED"
}

export enum PaymentStatus {
  UPCOMING = "UPCOMING",
  PENDING = "PENDING",
  COMPLETED = "COMPLETED"
}